feat(utils): calculate color's perceivede brightness to adjust genre colors
able to tell whether the theme is a dark or light theme.

adjusted genre-text-color to contrast the new genre-bg-color, meaning it
must be darker if the orignal color is light and vice versa. they both
come from the same base color; e.g. genre-text-color-00 and
genre-bg-color-00 both come from lets say a dark color-03 so text-color
is a lighter version of that, and bg-color is a darker version, creating
contrast.

new colors:
* genre-bg-color: a brighter (if already bright) or darker (if already
dark) version of color 03 to 05
* genre-outline-color: can be either color-00 or color-07, depending on
two things: whether the genre-bg-color is bright/dark AND whether the
current theme is a dark theme or a light theme.

these new colors ensure proper contrast in the smaller text of how
genres are presented

fix(templates): header navbar had wrong links