refactor(templates): heading surrounded by circles can pass in viewWidth and Height
This is for accurate width of svg for centering purposes. view height
must be equivalent with font-size for some reason. Just makes it so
sizing the actual svg heading is easier to wrap my head around. I
already did this for other headings; its just my first time writing
about it.

refactor(templates): use plain heading instead of followed by circles in
search results. not only because the circles are fixed in number and is
hard to adjust, but also mangagamer only has plain headings in search
results.