chore(Review, Comment): rename __ViewResponseDTOs to __ProfileViewResponseDTO since this is only used in the profile page
will name other response DTOs on where they are used e.g.
CommentReviewViewResponseDTO for comments that are under a review page
or something. This is different from the current
CommentProfileViewResponseDTO since in the profile view for the comments
section, I need review and book information, which I do not need in a
review page since the aforementioned review already has that.