feat(views): Book, Theme, and Review view services throw Exceptions over returning Result.Err
old returning Result.Err redirects to another page, not an error page.
There is one method left in Review that kept retuning a Result.Err:
invalid association. This is because while it is still an error, it is a
recoverable one, where we just need to redirect to the correct page with
the correct bookId associated with the review.