fix(templates): shown theme and used theme is not differentiated in theme preview
This means the unset button is always shown in the view themes page for
example. Theme preview works fine in all other pages (create theme,
confirm, and result) but not in view themes page because of the sidebar
right change. This fixes that by differentiating the shown and used
theme. the others can just pass in the same theme for shown and used